+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Json flatten VIEW example #65
 - Added CPU as user defined variable #67
+- Two target, for standalone and cluster version of presto #61
 
 ### Changed
 - Variables regarding credentials/secrets & updated documentation #52